pyproject.toml
[project]
name = "posebench"
version = "0.5.0"
description = "Comprehensive benchmarking of protein-ligand structure generation methods"
authors = [
    { name = "Alex Morehead", email = "acmwhb@umsystem.edu" }
]
readme = "README.md"
requires-python = ">= 3.7"
license = { file = "LICENSE" }
keywords = [
    'artificial intelligence',
    'deep learning',
    'protein-ligand structure generation',
    'benchmarking',
    'molecular docking',
    'protein-ligand docking',
    'protein-ligand interaction',
    'protein-ligand binding',
    'protein-ligand scoring',
    'protein-ligand pose prediction',
    'protein-ligand pose refinement',
]

classifiers=[
    'Development Status :: 4 - Beta',
    'Intended Audience :: Developers',
    'Topic :: Scientific/Engineering :: Artificial Intelligence',
    'License :: OSI Approved :: MIT License',
    'Programming Language :: Python :: 3.10',
]

dependencies = [
    # --------- pytorch --------- #
    "lightning",
    "torch",
    "torchmetrics",
    "torchvision",

    # --------- hydra --------- #
    "hydra-core==1.3.2",
    "hydra-colorlog==1.2.0",
    "hydra-optuna-sweeper==1.2.0",
    "omegaconf",
    "pydantic>=1.10.15",
    "pyyaml",

    # --------- others --------- #
    "beartype",        # runtime type checking
    "biopandas",       # for biological dataframe parsing
    "biopython==1.79", # for biological data parsing
    "ipykernel",       # for running Jupyter notebooks
    "jaxtyping>=0.2.12",  # tensor type checking
    "joblib",          # for parallel processing
    "lovely-numpy",    # for fancy numpy debugging
    "lovely-tensors",  # for fancy pytorch debugging
    "meeko",           # for molecular docking preparation
    "numpy",           # for numerical operations
    "pandas>=1.3.5",   # for data parsing and analysis
    "posebusters==0.2.12", # for protein-ligand structure scoring
    "plotly",          # for plotting
    "pre-commit",      # hooks for applying linters on commit
    "prody",           # for protein data parsing
    "prolif",          # for protein-ligand data parsing
    "pypdb",           # for PDB data parsing
    "rdkit>=2023.3.2", # for parsing small molecule data files
    "rich",            # beautiful text formatting in terminal
    "rootutils",       # standardizing the project root setup
    "scikit-learn>=1.0.2", # for basic machine learning utilities
    "seaborn",         # for plotting
    "setuptools",      # for packaging
    "spyrmsd",         # for RMSD calculations
    "timeout_decorator>=0.5.0", # for preventing certain functions from executing indefinitely
    "tqdm",            # progress bars
    "loguru",          # for cleaner logging
]

[project.urls]
Homepage = "https://pypi.org/project/posebench/"
Repository = "https://github.com/BioinfoMachineLearning/PoseBench"

[project.optional-dependencies]
examples = []
test = [
    "pytest",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = [
  "."
]
addopts = [
  "--color=yes",
  "--durations=0",
  "--strict-markers",
  "--doctest-modules",
]
filterwarnings = [
  "ignore::DeprecationWarning",
  "ignore::UserWarning",
]
log_cli = "True"
markers = [
  "slow: slow tests",
]
minversion = "6.0"
testpaths = "tests/"

# Assuming you're developing for Python 3.10
target-version = "py310"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["posebench"]

[tool.coverage.report]
exclude_lines = [
    "pragma: nocover",
    "raise NotImplementedError",
    "raise NotImplementedError()",
    "if __name__ == .__main__.:",
]
